Split servo code into servos.c
