+++ src/main/blackbox/blackbox.c
@@ -70,6 +70,7 @@
 #include "telemetry/telemetry.h"
 
 #include "flight/mixer.h"
+#include "flight/servos.h"
 #include "flight/failsafe.h"
 #include "flight/imu.h"
 #include "flight/navigation_rewrite.h"
