+++ src/main/blackbox/blackbox_io.c
@@ -67,6 +67,7 @@
 #include "common/printf.h"
 
 #include "flight/mixer.h"
+#include "flight/servos.h"
 #include "flight/failsafe.h"
 #include "flight/imu.h"
 #include "flight/navigation_rewrite.h"
