+++ src/main/config/config_master.h
@@ -80,6 +80,7 @@ typedef struct master_s {
 
     // mixer-related configuration
     mixerConfig_t mixerConfig;
+    servoConfig_t servoConfig;
 
 #ifdef GPS
     gpsConfig_t gpsConfig;
