+++ src/main/fc/msp_server_fc.c
@@ -75,6 +75,7 @@
 #include "sensors/gyro.h"
 
 #include "flight/mixer.h"
+#include "flight/servos.h"
 #include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/hil.h"
