+++ src/main/io/ledstrip.c
@@ -64,6 +64,7 @@
 
 #include "flight/failsafe.h"
 #include "flight/mixer.h"
+#include "flight/servos.h"
 #include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/navigation_rewrite.h"
