+++ src/main/telemetry/mavlink.c
@@ -59,6 +59,7 @@
 #include "rx/rx.h"
 
 #include "flight/mixer.h"
+#include "flight/servos.h"
 #include "flight/pid.h"
 #include "flight/imu.h"
 #include "flight/failsafe.h"
